Other changes in shared_view example
